fix: 没有更新服务器 & 运行库已存在时仍旧会尝试释放

+++ Plain Craft Launcher 2/Application.xaml.vb
@@ -143,12 +143,14 @@ WaitRetry:
             Directory.CreateDirectory(PathPure & "CE")
             SetDllDirectory(PathPure & "CE")
             WriteFile(PathPure & "CE\" & "libwebp.dll", GetResources("libwebp64"))
-            WriteFile(PathPure & "CE\" & "msalruntime.zip", GetResources("msalruntime"))
-            Using fs = New FileStream(PathPure & "CE\" & "msalruntime.zip", FileMode.Open)
-                Using fszip = New ZipArchive(fs)
-                    fszip.ExtractToDirectory(Path)
+            If Not Directory.Exists(Path & "runtimes") Then
+                WriteFile(PathPure & "CE\" & "msalruntime.zip", GetResources("msalruntime"))
+                Using fs = New FileStream(PathPure & "CE\" & "msalruntime.zip", FileMode.Open)
+                    Using fszip = New ZipArchive(fs)
+                        fszip.ExtractToDirectory(Path)
+                    End Using
                 End Using
-            End Using
+            End If
             '网络配置初始化
             ServicePointManager.Expect100Continue = True
             ServicePointManager.SecurityProtocol = SecurityProtocolType.Ssl3 Or SecurityProtocolType.Tls Or SecurityProtocolType.Tls11 Or SecurityProtocolType.Tls12
